+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryAndroidTest.kt
@@ -371,7 +371,7 @@ class SentryAndroidTest {
                 options.release = "prod"
                 options.dsn = "https://key@sentry.io/123"
                 options.isEnableAutoSessionTracking = true
-                options.experimental.sessionReplay.onErrorSampleRate = 1.0
+                options.sessionReplay.onErrorSampleRate = 1.0
                 optionsConfig(options)
             }