+++ sentry/src/main/java/io/sentry/ExperimentalOptions.java
@@ -1,7 +1,6 @@
 package io.sentry;
 
 import io.sentry.protocol.SdkVersion;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -11,18 +10,6 @@ import org.jetbrains.annotations.Nullable;
  * <p>Beware that experimental options can change at any time.
  */
 public final class ExperimentalOptions {
-  private @NotNull SentryReplayOptions sessionReplay;
 
-  public ExperimentalOptions(final boolean empty, final @Nullable SdkVersion sdkVersion) {
-    this.sessionReplay = new SentryReplayOptions(empty, sdkVersion);
-  }
-
-  @NotNull
-  public SentryReplayOptions getSessionReplay() {
-    return sessionReplay;
-  }
-
-  public void setSessionReplay(final @NotNull SentryReplayOptions sessionReplayOptions) {
-    this.sessionReplay = sessionReplayOptions;
-  }
+  public ExperimentalOptions(final boolean empty, final @Nullable SdkVersion sdkVersion) {}
 }